+++ app/src/main/kotlin/extraFunction/Digitizer.kt
@@ -1,3 +1,10 @@
+package extraFunction
+
+import library.Book
+import library.Disc
+import library.Library
+import library.Newspaper
+
 class Digitizer{
     fun <T : Library> digitize(item: T): Disc {
         return when(item) {
