+++ app/src/main/kotlin/extraFunction/ToSelectedType.kt
@@ -1,3 +1,5 @@
+package extraFunction
+
 inline fun <T, reified Result> toType(list: List<T>): MutableList<Result> {
     val res = mutableListOf<Result>()
     for (i in list) {
