+++ app/src/main/kotlin/forManaging/Manager.kt
@@ -1,3 +1,4 @@
+package forManaging
 
 class Manager {
     fun <T> buy(shops: Shops<T>): T{
