+++ app/src/main/kotlin/library/InterfacesForLibrary.kt
@@ -1,3 +1,5 @@
+package library
+
 interface Readable {
     fun read()
 }
