+++ app/src/main/kotlin/library/Library.kt
@@ -1,0 +1,10 @@
+package library
+
+abstract class Library(val id: Int, var isAvailable: Boolean, val name: String): Returnable,
+    Readable, Takeable {
+    abstract  fun getDetailedInformation()
+
+    fun getSummaryInformation(): String {
+        return "$name доступна: ${if (isAvailable) "Да" else "Нет" }"
+    }
+}
